Associates controller renamed.
