+++ Clienting.Domain/Associate/Events/AssociateStartedWorkingAtClient.cs
