+++ Clienting.Domain/Associate/IAssociateRepository.cs
