+++ Clienting.Domain/Client/IClientRepository.cs
