+++ ElektraLabManagement.sln
@@ -3,15 +3,17 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.27130.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Clienting", "Clienting", "{44B710E1-2341-4E5D-B7F7-1C95DBB252A3}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{44B710E1-2341-4E5D-B7F7-1C95DBB252A3}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Clienting.Domain", "Clienting.Domain\Clienting.Domain.csproj", "{DADE3707-BAD1-40CA-A461-DD9D474F8B48}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Clienting.Infrustructure", "Clienting.Infrustructure\Clienting.Infrustructure.csproj", "{B8655F85-64B2-421C-A921-B7754968566D}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Clienting.Api", "Clienting.Api\Clienting.Api.csproj", "{14234A2D-7515-4BE7-A2AB-08A06846C0CB}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Clienting.Api", "Clienting.Api\Clienting.Api.csproj", "{14234A2D-7515-4BE7-A2AB-08A06846C0CB}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Clienting.UnitTest", "Clienting.UnitTest\Clienting.UnitTest.csproj", "{942CAFAC-30C9-48AA-8E7B-06C066ABCD1A}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{5988C881-E0C4-4F95-AE7D-1E2AFD579BC2}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -43,7 +45,7 @@ Global
 		{DADE3707-BAD1-40CA-A461-DD9D474F8B48} = {44B710E1-2341-4E5D-B7F7-1C95DBB252A3}
 		{B8655F85-64B2-421C-A921-B7754968566D} = {44B710E1-2341-4E5D-B7F7-1C95DBB252A3}
 		{14234A2D-7515-4BE7-A2AB-08A06846C0CB} = {44B710E1-2341-4E5D-B7F7-1C95DBB252A3}
-		{942CAFAC-30C9-48AA-8E7B-06C066ABCD1A} = {44B710E1-2341-4E5D-B7F7-1C95DBB252A3}
+		{942CAFAC-30C9-48AA-8E7B-06C066ABCD1A} = {5988C881-E0C4-4F95-AE7D-1E2AFD579BC2}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {BA5F37B0-23B3-4818-B9AA-E9AE4BE42271}
